Receipt and Product
totalBill data type was changed from Integer to Double. GetAllReceiptsUseCase and UpdateProductUseCase were created

+++ src/main/java/com/sofkaU/software/demo/collection/Bill.java
@@ -25,7 +25,7 @@ public class Bill {
     private String salesPerson;
 
     @NotBlank(message = "Total bill cannot be blank")
-    private Integer totalBill;
+    private Double totalBill;
 
     private List<Product> products;
